refactored sniffer methods to separated files
