+++ include/bluetooth.hpp
@@ -7,6 +7,7 @@
 #include <GUILib.hpp>
 #include <Sensors.hpp>
 #include <wifi.hpp>
+#include <sniffer.hpp>
 
 #define SERVICE_UUID        "c8d1d262-861f-4082-947e-f383a259aaf3"
 #define CHARAC_DATA_UUID    "b0f332a8-a5aa-4f3f-bb43-f99e7791ae01"
