+++ include/sniffer.hpp
@@ -1,0 +1,8 @@
+#include <wifi.hpp>
+#include <MACPool.hpp>
+#include <ConfigApp.hpp>
+
+void snifferInit();
+void snifferStop();
+void snifferLoop();
+uint16_t getPaxCount();
